+++ src/OpenTelemetry.AutoInstrumentation/Configurations/FileBasedConfiguration/YamlConfiguration.cs
@@ -22,4 +22,31 @@ internal class YamlConfiguration
     /// </summary>
     [YamlMember(Alias = "resource")]
     public ResourceConfiguration? Resource { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the SDK is disabled.
+    /// If omitted or null, false is used.
+    /// </summary>
+    [YamlMember(Alias = "disabled")]
+    public bool Disabled { get; set; } = false;
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the Fail Fast is enabled.
+    /// If omitted or null, false is used.
+    /// </summary>
+    [YamlMember(Alias = "fail_fast")]
+    public bool FailFast { get; set; } = false;
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the Flush On Unhandled Exception is enabled.
+    /// If omitted or null, false is used.
+    /// </summary>
+    [YamlMember(Alias = "flush_on_unhandled_exception")]
+    public bool FlushOnUnhandledException { get; set; } = false;
+
+    /// <summary>
+    /// Gets or sets the no-code tracing configuration.
+    /// </summary>
+    [YamlMember(Alias = "no_code/development")]
+    public NoCodeConfiguration? NoCode { get; set; }
 }